Update session details for 20 November 2025
email from 30 October 2025:

Dear HPC SIG,

I have asked some of you, and a few have given their feedback (thanks!) on the   Technological Expertise Nomenclature I started a few weeks ago. Some motivation and ideas on how to use it can be found   here.

Could you, please, (especially if you haven’t done so before) have a look at the Computing part of the nomenclature and leave your comments in the relevant cells (please, don’t edit the cells themselves)?

A specific question I have is, how to name the “Accelerated computing” category? In the last Skills survey, there are two overlapping (or rather one is a subset of the other) names, “Accelerated computing” and “GPU computing”, and I prefer not to have overlaps if that can be avoided 😉:
•	How shall we cover this expertise- just “Accelerated computing”; “AC” with GPU, DPU, etc. In brackets after; separate categories for “GPU computing” , “DPU computing”, etc. (which ones?); other?

Thanks!
Elena